requirements.txt
asgiref==3.5.0
Django==3.2
django-extensions==3.1.5
django-filter==21.1
django-utils-six==2.0
djangorestframework==3.13.1
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.1.0
Pillow==9.1.0
psycopg2==2.9.3
PyJWT==1.7.1
python-dotenv==0.20.0
pytz==2022.1
sqlparse==0.4.2
